+++ include/openrave/interface.h
@@ -169,6 +169,10 @@ private:
     /// Write the help commands to an output stream
     virtual bool _GetCommandHelp(std::ostream& sout, std::istream& sinput) const;
 
+    inline InterfaceBase& operator=(const InterfaceBase&r) {
+        throw openrave_exception("InterfaceBase copying not allowed");
+    }
+
     mutable boost::mutex _mutexInterface; ///< internal mutex for protecting data from methods that might be access from any thread (those methods should be commented).
     InterfaceType __type; ///< \see GetInterfaceType
     UserDataPtr __plugin; ///< handle to plugin that controls the executable code. As long as this plugin pointer is present, module will not be unloaded.